refactor: close #643 use react router
