+++ app/utils.ts
@@ -49,7 +49,7 @@ export function isIOS() {
 }
 
 export function useMobileScreen() {
-  const [isMobileScreen_, setIsMobileScreen] = useState(false);
+  const [isMobileScreen_, setIsMobileScreen] = useState(isMobileScreen());
   useEffect(() => {
     const onResize = () => {
       setIsMobileScreen(isMobileScreen());
